move core events into hook
