--- src/context/CoreEventContext.tsx
@@ -1,173 +1,0 @@
-import { invoke } from "@tauri-apps/api";
-import { listen } from "@tauri-apps/api/event";
-import { createContext, useContext, useEffect, useState } from "react";
-import {
-  CORE_EVENT,
-  DISABLE,
-  ENABLE,
-  GET_CONFIG,
-  SCAN_LENGTH,
-  START_SCAN,
-  STOP_SCAN,
-} from "../data/constants";
-import { assertExhaustive } from "../utils";
-import type { ReactNode } from "react";
-import type { FeCoreEvent } from "../../src-tauri/bindings/FeCoreEvent";
-import type { FeVibeCheckConfig } from "../../src-tauri/bindings/FeVibeCheckConfig";
-import { FeStateEvent } from "../../src-tauri/bindings/FeStateEvent";
-import { createToast } from "../components/Toast";
-
-type CoreContextProps = {
-  isScanning: boolean;
-  isEnabled: boolean;
-  toggleIsEnabled: () => void;
-  toggleScan: () => void;
-  config: FeVibeCheckConfig | null;
-  refreshConfig: () => void;
-};
-const INITIAL_CORE_STATE: CoreContextProps = {
-  isScanning: false,
-  isEnabled: false,
-  toggleIsEnabled: () => null,
-  toggleScan: () => null,
-  config: null,
-  refreshConfig: () => null,
-};
-const CoreEventContext = createContext<CoreContextProps>(INITIAL_CORE_STATE);
-
-export function useCoreEventContext() {
-  return useContext(CoreEventContext);
-}
-
-export function CoreEventProvider({ children }: { children: ReactNode }) {
-  const [isEnabled, setIsEnabled] = useState(INITIAL_CORE_STATE.isEnabled);
-  const [isScanning, setIsScanning] = useState(INITIAL_CORE_STATE.isScanning);
-
-  const [config, setConfig] = useState<FeVibeCheckConfig | null>(
-    INITIAL_CORE_STATE.config
-  );
-
-  async function enable() {
-    try {
-      await invoke(ENABLE);
-      setIsEnabled(true);
-    } catch (e) {
-      createToast("error", "Could not enable!", JSON.stringify(e));
-    }
-  }
-
-  async function stopScanAndDisable() {
-    try {
-      await stopScan();
-      await invoke(DISABLE);
-      setIsEnabled(false);
-    } catch (e) {
-      createToast("error", "Could not disable!", JSON.stringify(e));
-    }
-  }
-
-  async function toggleIsEnabled() {
-    isEnabled ? stopScanAndDisable() : enable();
-  }
-
-  async function enableAndStartScan() {
-    try {
-      await enable();
-      await invoke(START_SCAN);
-      setIsScanning(true);
-    } catch (e) {
-      createToast("error", "Could not start scan!", JSON.stringify(e));
-    }
-  }
-
-  async function stopScan() {
-    try {
-      await invoke(STOP_SCAN);
-      setIsScanning(false);
-    } catch (e) {
-      createToast("error", "Could not stop scan!", JSON.stringify(e));
-    }
-  }
-
-  function toggleScan() {
-    isScanning ? stopScan() : enableAndStartScan();
-  }
-
-  useEffect(() => {
-    if (!isScanning) return;
-    const i = setInterval(() => stopScan(), SCAN_LENGTH);
-    return () => clearInterval(i);
-  }, [isScanning]);
-
-  function handleStateEvent(payload: FeStateEvent) {
-    switch (payload) {
-      case "Disable":
-        stopScanAndDisable();
-        break;
-      case "EnableAndScan":
-        enableAndStartScan();
-        break;
-      default:
-        assertExhaustive(payload);
-    }
-  }
-
-  function handleCoreEvent(payload: FeCoreEvent) {
-    switch (payload.kind) {
-      case "Scan":
-        setIsScanning(payload.data == "Start");
-        break;
-      case "State":
-        handleStateEvent(payload.data);
-        break;
-      default:
-        assertExhaustive(payload);
-    }
-  }
-
-  useEffect(() => {
-    const unlistenPromise = listen<FeCoreEvent>(CORE_EVENT, (event) =>
-      handleCoreEvent(event.payload)
-    );
-
-    return () => {
-      unlistenPromise.then((unlisten) => unlisten());
-    };
-  }, []);
-
-  async function refreshConfig() {
-    try {
-      const config = await invoke<FeVibeCheckConfig>(GET_CONFIG);
-      setConfig(config);
-    } catch {
-      setConfig(null);
-    }
-  }
-
-  useEffect(() => {
-    async function getConfig() {
-      try {
-        const config = await invoke<FeVibeCheckConfig>(GET_CONFIG);
-        setConfig(config);
-      } catch {
-        setConfig(null);
-      }
-    }
-    getConfig();
-  }, []);
-
-  return (
-    <CoreEventContext.Provider
-      value={{
-        isScanning,
-        isEnabled,
-        toggleIsEnabled,
-        toggleScan,
-        config,
-        refreshConfig,
-      }}
-    >
-      {children}
-    </CoreEventContext.Provider>
-  );
-}
